Fix: math functions on VS2010
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@8532 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/Handle Items.cpp
@@ -9879,7 +9879,7 @@ void TakePhoto(SOLDIERTYPE* pSoldier, INT32 sGridNo, INT8 bLevel )
 
 	for ( int x = -radius; x < radius; ++x )
 	{
-		int diff = sqrt( radius*radius - x*x );
+		int diff = sqrt( (FLOAT)(radius*radius - x*x) );
 
 		for ( int y = -diff; y < diff; ++y )
 		{
